wallet

+++ react-app/src/components/Wallet/index.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Redirect } from "react-router-dom";
 import UpdateWalletModal from "../UpdateWalletModal"
+import PostWalletModal from "../PostWalletModal";
 import OpenModalButton from "../OpenModalButton";
 import { fetchWallet, destroyWallet, updateUserWallet } from "../../store/wallet";
 import './wallet.css'
@@ -14,7 +15,13 @@ const Wallet = () => {
     const [isLoaded, setIsLoaded] = useState(false);
 
     const deleteWallet = (walletId) => {
-        dispatch(destroyWallet(walletId))
+        const confirm = window.confirm(
+            'Are you sure? This action cannot be undone'
+        )
+        if(confirm){
+            dispatch(destroyWallet(walletId))
+        }
+        
     }
 
     const fetchUpdateWallet = async () => {
@@ -29,24 +36,48 @@ const Wallet = () => {
         dispatch(fetchWallet()).then(() => setIsLoaded(true));
     }, [dispatch]);
 
+    const isWalletsEmpty = Object.keys(allWallets).length === 0;
+
     return (
         <div>
-            {Object.values(allWallets).map(wallet => (
+          {!isWalletsEmpty ? (
+           
+            <>
+             {console.log(allWallets, '!!!!!!!!!!!!!!!!')}
+              {Object.values(allWallets).map((wallet) => (
                 <div key={wallet.id}>
-                    <p>{wallet.account_type}</p>
-                    <p>{wallet.routing_num}</p>
-                    <p>{wallet.cash}</p>
-                    <button onClick={() => deleteWallet(wallet.id)}>Delete Wallet</button>
-                    {/* <UpdateWalletModal wallet={wallet}/> */}
-                    <OpenModalButton
-                        buttonText="Update Wallet"
-                        modalComponent={<UpdateWalletModal wallet={wallet} fetchUpdateWallet={fetchUpdateWallet}/>}
-                    />
+                  <p>{wallet.account_type}</p>
+                  <p>{wallet.routing_num}</p>
+                  <p>{wallet.cash}</p>
+                  <button onClick={() => deleteWallet(wallet.id)}>
+                    Delete Wallet
+                  </button>
+                  <OpenModalButton
+                    buttonText="Update Wallet"
+                    modalComponent={
+                      <UpdateWalletModal
+                        wallet={wallet}
+                        fetchUpdateWallet={fetchUpdateWallet}
+                      />
+                    }
+                  />
                 </div>
-            ))}
-
+              ))}
+            </>
+          ) : (
+            <div>
+            <p>Marlon is Gay</p>
+            <OpenModalButton
+              buttonText="Create Wallet"
+              modalComponent={
+                <PostWalletModal fetchUpdateWallet={fetchUpdateWallet} />
+                
+              }
+            />
+            </div>
+          )}
         </div>
-    );
-}
+      );
+            }
 
 export default Wallet;
